iox-#1640 Move PolymorphicSingleton impl
