+++ iceoryx_hoofs/include/iceoryx_hoofs/design_pattern/polymorphic_handler.hpp
@@ -1,44 +1,52 @@
-#pragma once
+// Copyright (c) 2022 by Apex.AI Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef IOX_HOOFS_DESIGN_PATTERN_POLYMORPHIC_HANDLER_HPP
+#define IOX_HOOFS_DESIGN_PATTERN_POLYMORPHIC_HANDLER_HPP
 
 #include <atomic>
 #include <iostream>
 #include <mutex>
 #include <type_traits>
 
-// TODO: change namespace? but DesignPattern is a bad namespace name ...
 namespace iox
 {
-namespace design
+/// @todo: change namespace name for consistency?
+namespace design_pattern
 {
 
-// TODO: namespace of this concept? should not depend on templated class (e.g. PolymorphicHandler)
-// it is not internal as it is supposed to be used to design new Handlers
-
 /// @brief Implements the Activatable concept to be used in the PolymorphicHandler
-///        The concept implements a binary switch. By default is switched on (active).
+/// The concept implements a binary switch. By default is switched on (active).
+/// Anyone defining another custom handler interface is supposed to derive from Activatable.
+/// @note While this is public, it is also partially implementation detail and partially convenience
+/// to use the PolymorphicHandler.
 class Activatable
 {
   public:
     Activatable() = default;
 
     /// @brief Switch on.
-    void activate()
-    {
-        m_active = true;
-    }
+    void activate() noexcept;
 
     /// @brief Switch off.
-    void deactivate()
-    {
-        m_active = false;
-    }
+    void deactivate() noexcept;
 
     /// @brief Query switch state.
     /// @return true if active (on), false otherwise (off).
-    bool isActive() const
-    {
-        return m_active;
-    }
+    bool isActive() const noexcept;
 
   private:
     bool m_active{true};
@@ -53,11 +61,9 @@ template <typename Interface>
 struct DefaultHooks
 {
     /// @brief called after if the polymorphic handler is set or reset after finalize
-    static void onSetAfterFinalize(Interface& /*currentInstance*/, Interface& /*newInstance*/)
-    {
-        std::cerr << "setting the polymorphic handler after finalize is not allowed" << std::endl;
-        std::terminate();
-    }
+    /// @param currentInstance the current instance of the handler singleton
+    /// @param newInstance the instance of the handler singleton to be set
+    static void onSetAfterFinalize(Interface& currentInstance, Interface& newInstance) noexcept;
 };
 
 } // namespace detail
@@ -88,136 +94,38 @@ class PolymorphicHandler
   public:
     /// @brief get the current singleton instance
     /// @return the current instance
-    static Interface& get();
+    static Interface& get() noexcept;
 
     /// @brief set the current singleton instance
     /// @param handler the handler instance to be set
     /// @return pointer to the previous instance
-    static Interface* set(Interface& handler);
+    static Interface* set(Interface& handler) noexcept;
 
     /// @brief reset the current singleton instance to the default instance
     /// @return pointer to the previous instance
-    static Interface* reset();
+    static Interface* reset() noexcept;
 
     /// @brief finalizes the instance, afterwards Hooks::onSetAfterFinalize
     ///        will be called during the remaining program lifetime
-    static void finalize();
+    static void finalize() noexcept;
 
   private:
     std::atomic_bool m_isFinal{false};
     std::mutex m_mutex;
     std::atomic<Interface*> m_current;
 
-    PolymorphicHandler();
+    PolymorphicHandler() noexcept;
 
-    static PolymorphicHandler& instance();
+    static PolymorphicHandler& instance() noexcept;
 
-    static Default& getDefault();
+    static Default& getDefault() noexcept;
 
-    static Interface* getCurrent();
+    static Interface* getCurrent() noexcept;
 };
 
-// on first call (in a thread):
-// 1. localHandler is initialized
-//    - getCurrent is called
-//    - instantiates singleton instance()
-//    - instantiates default
-//    - sets m_current of instance to default instance (release)
-//    - default is active
-
-// if any thread changes the active handler with set (or reset)
-// under mutex protection, it will:
-//    - set the new handler to active
-//    - set current handler to the new handler
-//    - deactivate the old handler (can still be used as it still needs to exist)
-//
-
-// on any call after the handler was changes in another thread
-// 1. we check whether the handler is active
-// (can be outdated information but will eventually be false once the atomic value is updated)
-// 2. if it was changed it is now inactive and we wait to obtain the mutex lock
-// under lock, we update the local handler to the new one (note that it cannot change
-// while this happens as we hold the lock)
-
-template <typename Interface, typename Default, typename Hooks>
-Interface& PolymorphicHandler<Interface, Default, Hooks>::get()
-{
-    // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables) false positive, thread_local
-    thread_local Interface* localHandler = getCurrent(); // initialized once per thread on first call
-
-    if (!localHandler->isActive())
-    {
-        std::lock_guard<std::mutex> lock(instance().m_mutex);
-        localHandler = getCurrent();
-    }
-    return *localHandler;
-}
-
-template <typename Interface, typename Default, typename Hooks>
-Interface* PolymorphicHandler<Interface, Default, Hooks>::set(Interface& handler)
-{
-    auto& ins = instance();
-
-    // setting is rare and we lock to synchronize and update the active flags properly
-    std::lock_guard<std::mutex> lock(ins.m_mutex);
-    if (ins.m_isFinal)
-    {
-        Hooks::onSetAfterFinalize(*getCurrent(), handler);
-        return nullptr;
-    }
-
-    handler.activate(); // it may have been deactivated before, so always reactivate it
-    auto prev = ins.m_current.exchange(&handler, std::memory_order_relaxed);
-
-    // anyone still using it will eventually see that it is inactive
-    // and switch to the new handler
-    prev->deactivate();
-    return prev;
-}
-
-template <typename Interface, typename Default, typename Hooks>
-Interface* PolymorphicHandler<Interface, Default, Hooks>::reset()
-{
-    return set(getDefault());
-}
-
-template <typename Interface, typename Default, typename Hooks>
-void PolymorphicHandler<Interface, Default, Hooks>::finalize()
-{
-    auto& ins = instance();
-    std::lock_guard<std::mutex> lock(ins.m_mutex);
-    ins.m_isFinal.store(true);
-}
-
-template <typename Interface, typename Default, typename Hooks>
-PolymorphicHandler<Interface, Default, Hooks>::PolymorphicHandler()
-{
-    std::lock_guard<std::mutex> lock(m_mutex);
-    //  this is set the first time we call instance()
-    m_current.store(&getDefault(), std::memory_order_release);
-}
-
-template <typename Interface, typename Default, typename Hooks>
-PolymorphicHandler<Interface, Default, Hooks>& PolymorphicHandler<Interface, Default, Hooks>::instance()
-{
-    static PolymorphicHandler handler;
-    return handler;
-}
-
-template <typename Interface, typename Default, typename Hooks>
-Default& PolymorphicHandler<Interface, Default, Hooks>::getDefault()
-{
-    static Default defaultHandler;
-    return defaultHandler;
-}
-
-template <typename Interface, typename Default, typename Hooks>
-Interface* PolymorphicHandler<Interface, Default, Hooks>::getCurrent()
-{
-    auto& ins = instance();
-    return ins.m_current.load(
-        std::memory_order_acquire); // must be strong enough to sync memory of the object pointed to
-}
-
-} // namespace design
+} // namespace design_pattern
 } // namespace iox
+
+#include "iceoryx_hoofs/internal/design_pattern/polymorphic_handler.inl"
+
+#endif // IOX_HOOFS_DESIGN_PATTERN_POLYMORPHIC_HANDLER_HPP
